"Implement basic data structures in Python and C++"
Added implementations of Stack and Queue data structures, in both C++ and Python. This includes operations like push, pop, top, front, back, size, and is_empty.

Each of these data structures and operations are necessary building blocks for a wide variety of algorithms and data manipulations in coding problems. The goal of this commit is to provide reusable and reliable implementations of these fundamental data structures. The CMakeLists.txt was also adjusted to accommodate the new files.