"Change of project name and addition of new problem solution"
Renamed project from 'test' to '12865' and changed the source file name accordingly in CMakeLists.txt. This is to keep the project name consistent with the problem number from BaekJoon Platform. Furthermore, added a new solution for problem '12865' in C++. This solution uses dynamic programming to solve this problem efficiently and contains reusable functions for maximizing a value.