Add new Python and C++ solutions and CMakeLists
Added new algorithms in Python language to solve problems in level 2, 3 and 4 of Goorm platform. Also added a C++ solution to a level 2 problem. Added a CMakeLists.txt file to manage the C++ build system. These updates provide solutions to various algorithm problems and a build system to compile and run the C++ codes efficiently.